@@ -1,40 +1,40 @@
-go 1.22
-
-toolchain go1.22.5
-
 module github.com/ory/oathkeeper
 
-replace github.com/mattn/go-sqlite3 => github.com/mattn/go-sqlite3 v1.14.10
+go 1.25
+
+replace github.com/mattn/go-sqlite3 => github.com/mattn/go-sqlite3 v1.14.28
+
+replace github.com/ory/x => ./oryx
 
 require (
 	github.com/Azure/azure-pipeline-go v0.2.2
 	github.com/Azure/azure-storage-blob-go v0.9.0
-	github.com/Masterminds/sprig/v3 v3.2.2
-	github.com/auth0/go-jwt-middleware/v2 v2.2.2
-	github.com/aws/aws-sdk-go v1.55.5
+	github.com/Masterminds/sprig/v3 v3.3.0
+	github.com/auth0/go-jwt-middleware/v2 v2.3.0
+	github.com/aws/aws-sdk-go v1.55.6
 	github.com/blang/semver v3.5.1+incompatible
 	github.com/dgraph-io/ristretto v1.0.0
 	github.com/dlclark/regexp2 v1.2.0
 	github.com/ghodss/yaml v1.0.0
 	github.com/go-faker/faker/v4 v4.0.0-beta.2
-	github.com/go-jose/go-jose/v3 v3.0.3
-	github.com/go-openapi/errors v0.20.3
-	github.com/go-openapi/runtime v0.24.2
-	github.com/go-openapi/strfmt v0.21.3
-	github.com/go-openapi/swag v0.22.3
-	github.com/go-openapi/validate v0.22.0
-	github.com/go-sql-driver/mysql v1.8.1
+	github.com/go-jose/go-jose/v3 v3.0.4
+	github.com/go-openapi/errors v0.22.2
+	github.com/go-openapi/runtime v0.28.0
+	github.com/go-openapi/strfmt v0.23.0
+	github.com/go-openapi/swag v0.23.1
+	github.com/go-openapi/validate v0.24.0
+	github.com/go-sql-driver/mysql v1.9.3
 	github.com/go-swagger/go-swagger v0.30.0
 	github.com/gobuffalo/httptest v1.5.2
 	github.com/gobwas/glob v0.2.3
-	github.com/golang-jwt/jwt/v5 v5.2.1
+	github.com/gofrs/uuid v4.4.0+incompatible
+	github.com/golang-jwt/jwt/v5 v5.3.0
 	github.com/golang/gddo v0.0.0-20190904175337-72a348e765d2
 	github.com/golang/mock v1.6.0
 	github.com/google/go-replayers/httpreplay v1.1.1
-	github.com/google/uuid v1.6.0
-	github.com/hashicorp/go-retryablehttp v0.7.7
+	github.com/hashicorp/go-retryablehttp v0.7.8
 	github.com/julienschmidt/httprouter v1.3.0
-	github.com/knadh/koanf/v2 v2.0.1
+	github.com/knadh/koanf/v2 v2.2.2
 	github.com/lib/pq v1.10.9
 	github.com/mitchellh/copystructure v1.2.0
 	github.com/ory/analytics-go/v5 v5.0.1
